requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
efficientnet-pytorch>=0.7.1
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
tqdm>=4.62.0
PyYAML>=5.4.0